rename : 기존 option을 shoe option으로 rename 및 기존 클래스 삭제
