+++ fullver/backend/src/main/java/com/github/chore/repository/entity/shoe_option/ShoeOption.java
@@ -1,4 +1,4 @@
-package com.github.chore.repository.entity.option;
+package com.github.chore.repository.entity.shoe_option;
 
 import com.github.chore.repository.entity.shoe.Shoe;
 import jakarta.persistence.*;
